refactor initialization of client using AppConfig
Since we dropped support for older Django versions, we can make use of
the new `AppConfig.ready` hook to initialize our handlers.

We also add automatic instrumentation of Celery tasks in the `ready`
handler.
