+++ elasticapm/contrib/celery/__init__.py
@@ -11,6 +11,8 @@ Large portions are
 
 from celery import signals
 
+from elasticapm.utils import get_name_from_func
+
 
 class CeleryFilter(object):
     def filter(self, record):
@@ -20,7 +22,7 @@ class CeleryFilter(object):
             return 1
 
 
-def register_signal(client):
+def register_exception_tracking(client):
     def process_failure_signal(sender, task_id, exception, args, kwargs,
                                traceback, einfo, **kw):
         client.capture_exception(
@@ -31,3 +33,14 @@ def register_signal(client):
                 'kwargs': kwargs,
             })
     signals.task_failure.connect(process_failure_signal, weak=False)
+
+
+def register_instrumentation(client):
+    def begin_transaction(*args, **kwargs):
+        client.begin_transaction("celery")
+
+    def end_transaction(task_id, task, *args, **kwargs):
+        name = get_name_from_func(task)
+        client.end_transaction(name, 200)
+    signals.task_prerun.connect(begin_transaction, weak=False)
+    signals.task_postrun.connect(end_transaction, weak=False)
